requirements.txt
# For installing everything at one run `pip install -r /path/to/requirements.txt`, where you of course have to fill in the path yourself.

altgraph==0.17
atomicwrites==1.4.0
attrs==20.3.0
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
colorama==0.4.4
coloredlogs==15.0
ffmpeg==1.4
ffmpeg-python==0.2.0
Flask==1.1.2
Flask-Cors==3.0.10
future==0.18.2
google-api-core==1.25.1
google-auth==1.25.0
google-cloud==0.34.0
google-cloud-core==1.5.0
google-cloud-speech==2.0.1
google-cloud-storage==1.35.0
google-cloud-translate==3.0.2
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
grpcio==1.35.0
humanfriendly==9.1
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
joblib==1.0.0
libcst==0.3.16
MarkupSafe==1.1.1
mypy-extensions==0.4.3
nltk==3.5
packaging==20.9
pefile==2019.4.18
pluggy==0.13.1
proto-plus==1.13.0
protobuf==3.14.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyinstaller==4.1
pyinstaller-hooks-contrib==2020.11
pyparsing==2.4.7
pyreadline==2.1
pyserial==3.5
pytest==6.2.2
pytz==2021.1
pywin32-ctypes==0.2.0
PyYAML==5.3.1
regex==2020.11.13
requests==2.25.1
rsa==4.7
six==1.15.0
termcolor==1.1.0
toml==0.10.2
tqdm==4.56.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.3
Werkzeug==1.0.1
